+++ src/app/compare/CompareItem.m.scss.d.ts
@@ -5,10 +5,7 @@ interface CssExports {
   'compareItem': string;
   'header': string;
   'highlightBar': string;
-  'initialItem': string;
-  'isFindable': string;
   'itemAside': string;
-  'itemName': string;
   'sortAsc': string;
   'sortDesc': string;
   'spacer': string;