+++ src/app/compare/CompareItem.tsx
@@ -1,4 +1,3 @@
-import BungieImage from 'app/dim-ui/BungieImage';
 import { PressTip } from 'app/dim-ui/PressTip';
 import { useDynamicStringReplacer } from 'app/dim-ui/destiny-symbols/RichDestinyText';
 import { ColumnSort, SortDirection } from 'app/dim-ui/table-columns';
@@ -9,13 +8,12 @@ import { currentStoreSelector, notesSelector } from 'app/inventory/selectors';
 import ActionButton from 'app/item-actions/ActionButton';
 import { LockActionButton, TagActionButton } from 'app/item-actions/ActionButtons';
 import { useD2Definitions } from 'app/manifest/selectors';
-import { statLabels } from 'app/organizer/Columns';
+import { ColumnDefinition, Row, TableContext } from 'app/organizer/table-types';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { noop } from 'app/utils/functions';
 import { useSetCSSVarToHeight, useShiftHeld } from 'app/utils/hooks';
 import { isD1Item } from 'app/utils/item-utils';
 import clsx from 'clsx';
-import { StatHashes } from 'data/d2/generated-enums';
 import { memo, useCallback, useMemo, useRef } from 'react';
 import { useSelector } from 'react-redux';
 import { useLocation } from 'react-router';
@@ -30,14 +28,13 @@ import {
   faArrowCircleDown,
   shoppingCart,
 } from '../shell/icons';
-import { StatInfo } from './Compare';
 import styles from './CompareItem.m.scss';
-import CompareStat from './CompareStat';
 
 export default memo(function CompareItem({
   item,
-  stats,
-  compareBaseStats,
+  row,
+  tableCtx,
+  filteredColumns,
   itemClick,
   remove,
   setHighlight,
@@ -45,8 +42,9 @@ export default memo(function CompareItem({
   isInitialItem,
 }: {
   item: DimItem;
-  stats: StatInfo[];
-  compareBaseStats?: boolean;
+  row: Row;
+  tableCtx: TableContext;
+  filteredColumns: ColumnDefinition[];
   itemClick: (item: DimItem) => void;
   remove: (item: DimItem) => void;
   setHighlight: (value?: string | number) => void;
@@ -80,15 +78,6 @@ export default memo(function CompareItem({
           {item.taggable ? <TagActionButton item={item} label={false} hideKeys={true} /> : <div />}
           <button type="button" className={styles.close} onClick={() => remove(item)} />
         </div>
-        <div
-          className={clsx(styles.itemName, {
-            [styles.initialItem]: isInitialItem,
-            [styles.isFindable]: isFindable,
-          })}
-          onClick={() => itemClick(item)}
-        >
-          <span title={isInitialItem ? t('Compare.InitialItem') : undefined}>{item.name}</span>
-        </div>
         <ItemPopupTrigger item={item} noCompare={true}>
           {(ref, onClick) => (
             <div className={styles.itemAside} ref={ref} onClick={onClick}>
@@ -100,7 +89,7 @@ export default memo(function CompareItem({
         </ItemPopupTrigger>
       </div>
     ),
-    [isInitialItem, item, itemClick, pullItem, remove, itemNotes, isFindable],
+    [item, pullItem, remove, itemNotes],
   );
 
   const missingSocketsMessage =
@@ -108,18 +97,28 @@ export default memo(function CompareItem({
       ? tl('MovePopup.MissingSockets')
       : tl('MovePopup.LoadingSockets');
 
+  const handleRowClick = (row: Row, column: ColumnDefinition) => {
+    if (column.id === 'name' && isFindable) {
+      return () => itemClick(row.item);
+    }
+    return undefined;
+  };
+
   return (
-    <div className={styles.compareItem}>
+    <div
+      className={clsx(styles.compareItem, {
+        'compare-initial': isInitialItem,
+        'compare-findable': isFindable,
+      })}
+    >
       {itemHeader}
-      {stats.map((stat) => (
-        <CompareStat
-          key={stat.stat.statHash}
-          item={item}
-          stat={stat}
-          setHighlight={setHighlight}
-          compareBaseStats={compareBaseStats}
-        />
-      ))}
+      <TableRow
+        row={row}
+        tableCtx={tableCtx}
+        filteredColumns={filteredColumns}
+        onRowClick={handleRowClick}
+        setHighlight={setHighlight}
+      />
       {isD1Item(item) && item.talentGrid && (
         <ItemTalentGrid item={item} className={styles.talentGrid} perksOnly={true} />
       )}
@@ -150,56 +149,97 @@ function VendorItemWarning({ item }: { item: DimItem }) {
   ) : null;
 }
 
+// Copied from ItemTable - TODO: reconverge
+function TableRow({
+  row,
+  tableCtx,
+  filteredColumns,
+  onRowClick,
+  setHighlight,
+}: {
+  row: Row;
+  tableCtx: TableContext;
+  filteredColumns: ColumnDefinition[];
+  onRowClick: (
+    row: Row,
+    column: ColumnDefinition,
+  ) => ((event: React.MouseEvent<HTMLTableCellElement>) => void) | undefined;
+  setHighlight: (value?: string | number) => void;
+}) {
+  return (
+    <>
+      {filteredColumns.map((column) => (
+        // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
+        <div
+          key={column.id}
+          onClick={onRowClick(row, column)}
+          className={clsx(column.className, {
+            // [styles.hasFilter]: column.filter !== undefined,
+          })}
+          role="cell"
+          onPointerEnter={() => setHighlight(column.id)}
+        >
+          {column.cell
+            ? column.cell(row.values[column.id], row.item, tableCtx.minMaxValues[column.id])
+            : row.values[column.id]}
+        </div>
+      ))}
+    </>
+  );
+}
+
 /** The row headers that appear on the left of the compare window */
 export function CompareHeaders({
   columnSorts,
   highlight,
   setHighlight,
   toggleColumnSort,
-  allStats,
+  filteredColumns,
 }: {
   columnSorts: ColumnSort[];
   highlight: string | number | undefined;
   setHighlight: React.Dispatch<React.SetStateAction<string | number | undefined>>;
   toggleColumnSort: (columnId: string, shiftHeld: boolean, sort?: SortDirection) => () => void;
-  allStats: StatInfo[];
+  filteredColumns: ColumnDefinition[];
 }) {
   const isShiftHeld = useShiftHeld();
   return (
     <div className={styles.statList}>
       <div className={styles.spacer} />
-      {allStats.map((s) => {
-        const columnSort = columnSorts.find((c) => c.columnId === s.stat.statHash.toString());
+      {filteredColumns.map((column) => {
+        const columnSort = !column.noSort && columnSorts.find((c) => c.columnId === column.id);
         return (
           <div
-            key={s.stat.statHash}
+            key={column.id}
             className={clsx(
               styles.statLabel,
+              column.headerClassName,
               columnSort
                 ? columnSort.sort === SortDirection.ASC
                   ? styles.sortDesc
                   : styles.sortAsc
                 : undefined,
             )}
-            onPointerEnter={() => setHighlight(s.stat.statHash)}
-            onClick={toggleColumnSort(
-              s.stat.statHash.toString(),
-              isShiftHeld,
-              s.stat.smallerIsBetter ? SortDirection.DESC : SortDirection.ASC,
-            )}
+            onPointerEnter={() => setHighlight(column.id)}
+            onClick={
+              column.noSort
+                ? undefined
+                : toggleColumnSort(column.id, isShiftHeld, column.defaultSort)
+            }
+            role="rowheader"
+            aria-sort={
+              columnSort
+                ? columnSort.sort === SortDirection.ASC
+                  ? 'ascending'
+                  : 'descending'
+                : 'none'
+            }
           >
-            {s.stat.displayProperties.hasIcon && (
-              <span title={s.stat.displayProperties.name}>
-                <BungieImage src={s.stat.displayProperties.icon} />
-              </span>
-            )}
-            {s.stat.statHash in statLabels
-              ? t(statLabels[s.stat.statHash as StatHashes]!)
-              : s.stat.displayProperties.name}{' '}
+            {column.header}{' '}
             {columnSort && (
               <AppIcon icon={columnSort.sort === SortDirection.ASC ? faAngleRight : faAngleLeft} />
             )}
-            {s.stat.statHash === highlight && <div className={styles.highlightBar} />}
+            {column.id === highlight && <div className={styles.highlightBar} />}
           </div>
         );
       })}