+++ src/app/organizer/Columns.tsx
@@ -82,6 +82,7 @@ import {
   csvStatNamesForDestinyVersion,
 } from 'app/inventory/spreadsheets';
 import { DeepsightHarmonizerIcon } from 'app/item-popup/DeepsightHarmonizerIcon';
+import { emptyArray } from 'app/utils/empty';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import styles from './Columns.m.scss';
 import { ColumnDefinition, ColumnGroup, ColumnWithStat, SortDirection, Value } from './table-types';
@@ -138,6 +139,7 @@ const perkStringFilter = (value: string | undefined) => {
 export function getColumns(
   useCase: 'organizer' | 'spreadsheet',
   itemsType: 'weapon' | 'armor' | 'ghost',
+  /** A single example stat per stat hash among items */
   stats: DimStat[],
   getTag: (item: DimItem) => TagValue | undefined,
   getNotes: (item: DimItem) => string | undefined,
@@ -266,6 +268,8 @@ export function getColumns(
               header: t('Organizer.Columns.StatQualityStat', {
                 stat: stat.displayProperties.name,
               }),
+              className: styles.stats,
+              headerClassName: styles.stats,
               value: (item: D1Item) => {
                 const stat = item.stats?.find((s) => s.statHash === statHash);
                 let pct = 0;
@@ -1144,9 +1148,11 @@ function getIntrinsicSockets(item: DimItem) {
  * It will return the same result for the same category, since all items in a category share stats.
  */
 export function buildStatInfo(items: DimItem[]): DimStat[] {
-  const statHashes: {
-    [statHash: number]: DimStat;
-  } = {};
+  if (!items.length) {
+    return emptyArray<DimStat>();
+  }
+
+  const statHashes: { [statHash: number]: DimStat } = {};
   for (const item of items) {
     if (item.stats) {
       for (const stat of item.stats) {