+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 8.63.0 <span class="changelog-date">(2025-03-16)</span>
 
 * Fixed a bug where the Organizer wouldn't show all the items if your screen was very tall or zoomed out.
+* Added a sortable name label to the Compare view.
 
 ## 8.62.0 <span class="changelog-date">(2025-03-09)</span>